pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amaya"
version = "0.2.0"
description = "ADI — AI Durability Index rating pipeline (scoring + ingest)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "YVL Capital Partners"}]
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5",
]

[project.scripts]
amaya = "amaya.cli:main"

[tool.setuptools.packages.find]
include = ["amaya*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
